Use config file to avoid many calls to `keybase`
Previously every command would shell out to the Keybase CLI tool to
determine the current authenticated user. This is slower than it needs
to be.

This commit adds a `$HOME/.passbase` - or `%USERPROFILE%\.passbase` if on
Windows - JSON configuration file with the structure:

    { "User": "keybase_user_name" }

which is used if available.

If the file does not exist, or does not contain the `"User"` key,
`keybase` will be invoked as previously and the file created or key
added as appropriate.

This commit closes #7.