Implement recovery of changed password
It's easy to imagine a case in which one decides to change a password,
navigates through a maze of settings on the website, and while doing so
runs `passbase change`.

Whoops.

Website asks for confirmation of existing password - it's gone. Perhaps
the browser's saved passwords feature can save the day, perhaps not. It
seems worthwhile having the ability to recover a previous password for
no real cost.

    passbase recover mytag

acts as `passbase read mytag`, but for the state of the password before
`password change mytag`. `passbase read` continues as normal.

Closes #11.